Move global definitions in JSignalHandler.h and JVersion.h to implementation files to avoid violating the One Definition Rule that was leading to double free seg. faults due to double free at program exit,
